Addressed PR comments, moved reused code into common location
