+++ src/snapred/ui/widget/SmoothingSlider.py
@@ -79,8 +79,9 @@ class SmoothingSlider(QWidget):
             )
             return
         s = float(text)
+        # Set v to -10 so that slider gets set to minimum
         if s == 0:
-            v = 0
+            v = -10
         else:
             v = math.log10(s)
         sliderValue = int(v * 100)
