+++ tests/integration/test_normalization.py
@@ -1,10 +1,7 @@
 import math
 import os
 import re
-import tempfile
 from contextlib import ExitStack, suppress
-from pathlib import Path
-from typing import Optional
 from unittest import mock
 
 import pytest
@@ -12,19 +9,14 @@ from mantid.kernel import amend_config
 from qtpy import QtCore
 from qtpy.QtCore import Qt
 from qtpy.QtWidgets import (
-    QMessageBox,
     QTabWidget,
 )
-from util.Config_helpers import Config_override
+from util.pytest_helpers import handleStateInit
 from util.qt_mock_util import MockQMessageBox
 from util.TestSummary import TestSummary
 
-# I would prefer not to access `LocalDataService` within an integration test,
-#   however, for the moment, the reduction-data output relocation fixture is defined in the current file.
-from snapred.backend.data.LocalDataService import LocalDataService
-from snapred.meta.Config import Config, Resource
+from snapred.meta.Config import Resource
 from snapred.ui.main import SNAPRedGUI, prependDataSearchDirectories
-from snapred.ui.view import InitializeStateCheckView
 from snapred.ui.view.NormalizationRequestView import NormalizationRequestView
 from snapred.ui.view.NormalizationSaveView import NormalizationSaveView
 from snapred.ui.view.NormalizationTweakPeakView import NormalizationTweakPeakView
@@ -34,56 +26,9 @@ class InterruptWithBlock(BaseException):
     pass
 
 
-@pytest.fixture
-def calibration_home_from_mirror():
-    # Test fixture to create a copy of the calibration home directory from an existing mirror:
-    # * creates a temporary calibration home directory under the optional `prefix` path;
-    #   when not specified, the temporary directory is created under the existing
-    #   `Config["instrument.calibration.powder.home"]`;
-    # * creates symlinks within the directory to required metadata files and directories
-    #   from the already existing `Config["instrument.calibration.powder.home"]`;
-    # * ignores any existing diffraction-calibration and normalization-calibration subdirectories;
-    # * and finally, overrides the `Config` entry for "instrument.calibration.powder.home".
-
-    # IMPLEMENTATION notes:
-    # * The functionality of this fixture is deliberately NOT implemented as a context manager,
-    #   although certain context-manager features are used.
-    # * If this were a context manager, it would  be terminated at any exception throw.  For example,
-    #   it would be terminated by the "initialize state" `RecoverableException`.  Such termination would interfere with
-    #   the requirements of the integration tests.
-    _stack = ExitStack()
-
-    def _calibration_home_from_mirror(prefix: Optional[Path] = None):
-        originalCalibrationHome: Path = Path(Config["instrument.calibration.powder.home"])
-        if prefix is None:
-            prefix = originalCalibrationHome
-
-        # Override calibration home directory:
-        tmpCalibrationHome = Path(_stack.enter_context(tempfile.TemporaryDirectory(dir=prefix, suffix=os.sep)))
-        assert tmpCalibrationHome.exists()
-        _stack.enter_context(Config_override("instrument.calibration.powder.home", str(tmpCalibrationHome)))
-
-        # WARNING: for these integration tests `LocalDataService` is a singleton.
-        #   The Indexer's `lru_cache` MUST be reset after the Config override, otherwise
-        #     it will return indexers synched to the previous `Config["instrument.calibration.powder.home"]`.
-        LocalDataService()._indexer.cache_clear()
-
-        # Create symlinks to metadata files and directories.
-        metadatas = [Path("LiteGroupMap.hdf"), Path("PixelGroupingDefinitions"), Path("SNAPLite.xml")]
-        for path_ in metadatas:
-            os.symlink(originalCalibrationHome / path_, tmpCalibrationHome / path_)
-        return tmpCalibrationHome
-
-    yield _calibration_home_from_mirror
-
-    # teardown => __exit__
-    _stack.close()
-    LocalDataService()._indexer.cache_clear()
-
-
 @pytest.mark.datarepo
 @pytest.mark.integration
-class TestGUIPanels:
+class TestNormalizationPanels:
     @pytest.fixture(scope="function", autouse=True)  # noqa: PT003
     def _setup_gui(self, qapp):
         testMock = MockQMessageBox()
@@ -95,7 +40,9 @@ class TestGUIPanels:
         # Automatically continue at the end of each workflow.
         self._actionPrompt = mock.patch(
             "qtpy.QtWidgets.QMessageBox.information",
-            lambda *args: TestGUIPanels._actionPromptContinue(*args, match=r".*has been completed successfully.*"),
+            lambda *args: TestNormalizationPanels._actionPromptContinue(
+                *args, match=r".*has been completed successfully.*"
+            ),
         )
         self._actionPrompt.start()
         # ---------------------------------------------------------------------------
@@ -259,7 +206,6 @@ class TestGUIPanels:
                 requestView.backgroundRunNumberField.setText("58813")
                 qtbot.mouseClick(workflowNodeTabs.currentWidget().continueButton, Qt.MouseButton.LeftButton)
                 qtbot.wait(100)
-
                 # Verify error msg box to select sample shows up
                 assert len(exceptions) == 0
                 assert mpCrit[1].call_count == 1
@@ -290,54 +236,8 @@ class TestGUIPanels:
             #   At the moment, we preserve some options here to speed up development.
             stateId = "04bd2c53f6bf6754"
             waitForStateInit = True
-            if (Path(Config["instrument.calibration.powder.home"]) / stateId).exists():
-                # raise RuntimeError(
-                #           f"The state root directory for '{stateId}' already exists! "\
-                #           + "Please move it out of the way."
-                #       )
-                waitForStateInit = False
 
-            if waitForStateInit:
-                # ---------------------------------------------------------------------------
-                # IMPORTANT: "initialize state" dialog is triggered by an exception throw:
-                #   => do _not_ patch using a with clause!
-                questionMessageBox = mock.patch(  # noqa: PT008
-                    "qtpy.QtWidgets.QMessageBox.question",
-                    lambda *args, **kwargs: QMessageBox.Yes,  # noqa: ARG005
-                )
-                questionMessageBox.start()
-                successPrompt = mock.patch(
-                    "snapred.ui.widget.SuccessPrompt.SuccessPrompt.prompt",
-                    lambda parent: parent.close() if parent is not None else None,
-                )
-                successPrompt.start()
-                # ---------------------------------------------------------------------------
-
-                #    (1) respond to the "initialize state" request
-                with qtbot.waitSignal(actionCompleted, timeout=60000):
-                    qtbot.mouseClick(workflowNodeTabs.currentWidget().continueButton, Qt.MouseButton.LeftButton)
-
-                qtbot.waitUntil(
-                    lambda: len(
-                        [
-                            o
-                            for o in qapp.topLevelWidgets()
-                            if isinstance(o, InitializeStateCheckView.InitializationMenu)
-                        ]
-                    )
-                    > 0,
-                    timeout=1000,
-                )
-                stateInitDialog = [
-                    o for o in qapp.topLevelWidgets() if isinstance(o, InitializeStateCheckView.InitializationMenu)
-                ][0]
-
-                stateInitDialog.stateNameField.setText("my happy state")
-                qtbot.mouseClick(stateInitDialog.beginFlowButton, Qt.MouseButton.LeftButton)
-
-                # State initialization dialog is "application modal" => no need to explicitly wait
-                questionMessageBox.stop()
-                successPrompt.stop()
+            handleStateInit(waitForStateInit, stateId, qtbot, qapp, actionCompleted, workflowNodeTabs)
 
             msg = "No valid FocusGroups were specified for mode: 'lite'"
             mp = MockQMessageBox().exec(msg)
@@ -358,7 +258,6 @@ class TestGUIPanels:
             msg = "Smoothing parameter must be a numerical value"
             mpWarn = MockQMessageBox().warning(msg)
             with mpWarn[0]:
-                mpWarn[0].start()
                 tweakPeakView.smoothingSlider.field.setValue("a")
                 qtbot.wait(100)
                 assert len(exceptions) == 0
@@ -374,7 +273,6 @@ class TestGUIPanels:
                 tweakPeakView.fieldXtalDMin.setText("a")
                 qtbot.mouseClick(tweakPeakView.recalculationButton, Qt.MouseButton.LeftButton)
                 qtbot.wait(100)
-                mpWarn[0].stop()
                 assert len(exceptions) == 0
                 assert mpWarn[1].call_count == 1
 
@@ -388,7 +286,6 @@ class TestGUIPanels:
                 tweakPeakView.fieldXtalDMax.setText("a")
                 qtbot.mouseClick(tweakPeakView.recalculationButton, Qt.MouseButton.LeftButton)
                 qtbot.wait(100)
-                mpWarn[0].stop()
                 assert len(exceptions) == 0
                 assert mpWarn[1].call_count == 1
 
@@ -400,7 +297,6 @@ class TestGUIPanels:
             with mpWarn[0]:
                 qtbot.mouseClick(tweakPeakView.recalculationButton, Qt.MouseButton.LeftButton)
                 qtbot.wait(500)
-                mpWarn[0].stop()
                 assert len(exceptions) == 0
                 assert mpWarn[1].call_count == 1
 
@@ -411,7 +307,6 @@ class TestGUIPanels:
                 tweakPeakView.fieldXtalDMax.setText("1")
                 tweakPeakView.smoothingSlider.field.setValue("-1")
                 qtbot.wait(100)
-                mpWarn[0].stop()
                 assert len(exceptions) == 0
                 assert mpWarn[1].call_count == 1
 
@@ -426,7 +321,6 @@ class TestGUIPanels:
                 tweakPeakView.fieldXtalDMin.setText("-1")
                 qtbot.mouseClick(tweakPeakView.recalculationButton, Qt.MouseButton.LeftButton)
                 qtbot.wait(100)
-                mpWarn[0].stop()
                 assert len(exceptions) == 0
                 assert mpWarn[1].call_count == 1
 
@@ -437,7 +331,6 @@ class TestGUIPanels:
                 tweakPeakView.fieldXtalDMax.setText("-1")
                 qtbot.mouseClick(tweakPeakView.recalculationButton, Qt.MouseButton.LeftButton)
                 qtbot.wait(100)
-                mpWarn[0].stop()
                 assert len(exceptions) == 0
                 assert mpWarn[1].call_count == 1
 
@@ -505,7 +398,7 @@ class TestGUIPanels:
                 qtbot.wait(100)
                 assert len(exceptions) == 0
                 assert mpCrit[1].call_count == 1
-                saveView.fieldAuthor.setText("Test")
+            saveView.fieldAuthor.setText("Test")
 
             # Set Comments to "This is a test"You must add comments
             msg = "You must add comments"
@@ -516,7 +409,7 @@ class TestGUIPanels:
                 mpCrit[0].stop()
                 assert len(exceptions) == 0
                 assert mpCrit[1].call_count == 1
-                saveView.fieldComments.setText("This is a test")
+            saveView.fieldComments.setText("This is a test")
 
             # Finish and observe "Workflow complete"
             #    continue in order to save workspaces and to finish the workflow
